+++ app/javascript/dashboard/components-next/Campaigns/CampaignCard/LiveChatCampaignDetails.vue
