+++ app/javascript/dashboard/components-next/Campaigns/CampaignCard/SMSCampaignDetails.vue
